Fulfill a Swap Request
Swap scheduled days for 2
employees. Create new schedule
objects with the correct employee/
date info, then delete the old
schedules in order to avoid validation
errors that would be encountered by
swaping in place.

Do this in a transaction so data
will be restored in the event of an
error.