Add Simple Authentication
Use current_user gem to add
very basic authentication, necessary
for adding conflicts, making swap
requests and confirming swap requests.

Create a User subclass of Employee
because the gem expects the users to be
stored in a model called User.

Override the header template of the gem
in order to customize it with a link
back to the schedule page.